# ProjectEshop.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ProjectEshop",
    "version": "v1"
  },
  "paths": {
...
    "/api/Score/ScoreGetAll": {
      "get": {
        "tags": [
          "Score"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Score/ScoreGetDetail": {
      "get": {
        "tags": [
          "Score"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Score/ScoreCreate": {
      "post": {
        "tags": [
          "Score"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Score/ScoreEdit": {
      "put": {
        "tags": [
          "Score"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Score/ScoreDelete": {
      "delete": {
        "tags": [
          "Score"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ManufacturerLogoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "logo": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "idManufacturer": {
            "type": "integer",
            "format": "int32"
          },
          "idManufacturerNavigation": {
            "$ref": "#/components/schemas/ManufacturerDto"
          }
        },
        "additionalProperties": false
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductImageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "picture": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "idProduct": {
            "type": "integer",
            "format": "int32"
          },
          "idProductNavigation": {
            "$ref": "#/components/schemas/ProductDto"
          }
        },
        "additionalProperties": false
      },
      "ScoreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "numberScore": {
            "type": "integer",
            "format": "int32"
          },
          "textScore": {
            "type": "string",
            "nullable": true
          },
          "idProduct": {
            "type": "integer",
            "format": "int32"
          },
          "idProductNavigation": {
            "$ref": "#/components/schemas/ProductDto"
          }
        },
        "additionalProperties": false
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "productName": {
            "type": "string",
            "nullable": true
          },
          "productDescription": {
            "type": "string",
            "nullable": true
          },
          "productPrice": {
            "type": "number",
            "format": "double"
          },
          "productWeight": {
            "type": "number",
            "format": "double"
          },
          "productQuantityInStock": {
            "type": "integer",
            "format": "int32"
          },
          "idCategory": {
            "type": "integer",
            "format": "int32"
          },
          "idManufacturer": {
            "type": "integer",
            "format": "int32"
          },
          "idCategoryNavigation": {
            "$ref": "#/components/schemas/CategoryDto"
          },
          "idManufacturerNavigation": {
            "$ref": "#/components/schemas/ManufacturerDto"
          },
          "productImages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductImageDto"
            },
            "nullable": true
          },
          "scores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ManufacturerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "manufacturerName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "manufacturerLogos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManufacturerLogoDto"
            },
            "nullable": true
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}